+++ src/modules/likes.js
@@ -1,7 +1,5 @@
 import axios from 'axios';
-
-const BaseURL = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps';
-const APP_ID = 'un4IMXxgYWzcZxzqDkPT';
+import { BASE_URL, APP_ID } from './utils.js';
 
 const addLike = async (itemID) => {
   if (itemID) {
@@ -10,14 +8,14 @@ const addLike = async (itemID) => {
       likes: 1,
     };
 
-    await axios.post(`${BaseURL}/${APP_ID}/likes`, newLike);
+    await axios.post(`${BASE_URL}/${APP_ID}/likes`, newLike);
     return true;
   }
   return false;
 };
 
 const getTotalLikes = async (itemID) => {
-  const response = await fetch(`${BaseURL}/${APP_ID}/likes`);
+  const response = await fetch(`${BASE_URL}/${APP_ID}/likes`);
   const data = await response.json();
   const results = data.find((item) => +item.item_id === +itemID);
   return results === undefined ? 0 : results.likes;
