+++ src/modules/utils.js
@@ -1,0 +1,7 @@
+const TV_API_URL = 'https://api.tvmaze.com';
+
+const BASE_URL =
+  'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps';
+const APP_ID = 'un4IMXxgYWzcZxzqDkPT';
+
+export { TV_API_URL, BASE_URL, APP_ID };
